src/requirements.txt
appdirs==1.4.4
asgiref==3.2.10
astroid==2.4.2
attrs==19.3.0
backcall==0.2.0
beautifulsoup4==4.9.3
black==19.10b0
click==7.1.2
decorator==4.4.2
Django==3.1
django-bootstrap4==2.2.0
django-filter==2.4.0
django-forms-bootstrap==3.1.0
django-tables2==2.3.2
importlib-metadata==1.7.0
ipython==7.16.1
ipython-genutils==0.2.0
isort==4.3.21
jedi==0.17.2
lazy-object-proxy==1.4.3
mccabe==0.6.1
mysql-connector-python==8.0.21
mysqlclient==2.0.1
parso==0.7.1
pathspec==0.8.0
pexpect==4.8.0
pickleshare==0.7.5
prompt-toolkit==3.0.7
protobuf==3.13.0
ptyprocess==0.6.0
Pygments==2.7.1
pylint==2.5.3
pytz==2020.1
regex==2020.7.14
six==1.15.0
soupsieve==2.0.1
sqlparse==0.3.1
toml==0.10.1
traitlets==4.3.3
typed-ast==1.4.1
wcwidth==0.2.5
wrapt==1.12.1
zipp==3.3.0
